Updated Social Media Icons
I removed the previous images with the social media icons and added new ones that comply with the guidelines set by YouTube, Instagram, Faccebook, and Twitter. I need to modify the code for the footer for the rest of the pages, since I only modified the sice of the icons in the index.html footer.