Added a White version of Wistarr's Logo
I added a white version of the company's logo. Also, I loaded the white version of the logo in both the footer and in the navbar while the navbar is transparent.But this is in index_play.html, NOT in the actual home page for Wistarr in index.html yet.