Updated Portfolio Detailed View Page
I moved the iframe with the video below the title that's below the huge thumbnail image. Also, I added the Police Tests' logo as the huge thumbnail for the portfolio detailed view page. Additionally, I added 2 buttons to the right of the iframes which will have the links to the repo and to the URL where the web app is deployed.