Current Page Reloads if Accepting Cookies
I the user clicks on "I accept" on the cookie consent banner, the current page reloads. Therefore, as soon as you click "Accept", Google Analytics will work, and the youtube videos in portfolio.html will load. You don't have to go to another page to finish loading the cookies and the youtube videos.