Updated Portfolio Overview Page
I added the links to the last 3 portfolio projects to their respective detailed view page. That is, if users click on any of the 4 portfolio projects, they will be redirected to the detailed view of that project. I also added the images for the big thumbnails for every project except for Eduardo Fox's Gallery.